Address routing review feedback
Align routing code with MEAI formatting and documentation conventions, simplify disposal, and split routing tests by component and assembly.

Copilot-Session: 74d04840-2379-4615-93f7-84f2299ada74
