+++ test/Libraries/Microsoft.Extensions.AI.Tests/ChatRouting/FailoverChatClientTests.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
@@ -11,141 +10,8 @@ using Xunit;
 
 namespace Microsoft.Extensions.AI;
 
-public class RoutingChatClientTests
+public class FailoverChatClientTests
 {
-    [Fact]
-    public void RoutingContext_CarriesRequestInputs()
-    {
-        var messages = new List<ChatMessage> { new(ChatRole.User, "initial") };
-        var options = new ChatOptions { ModelId = "initial" };
-        var context = new RoutingContext(messages, options);
-
-        Assert.Same(messages, context.Messages);
-        Assert.NotSame(options, context.ChatOptions);
-        Assert.Equal("initial", context.ChatOptions!.ModelId);
-        context.ChatOptions.ModelId = "changed";
-        Assert.Equal("initial", options.ModelId);
-        Assert.Throws<ArgumentNullException>(() => new RoutingContext(null!, options));
-    }
-
-    [Fact]
-    public void Create_RejectsNullSelector()
-    {
-        Assert.Throws<ArgumentNullException>(() => RoutingChatClient.Create(null!));
-    }
-
-    [Fact]
-    public async Task Create_SelectsClientForRequest()
-    {
-        var messages = new ChatMessage[] { new(ChatRole.User, "hi") };
-        var options = new ChatOptions { ModelId = "request" };
-        using var cancellationSource = new CancellationTokenSource();
-        ChatResponse expected = new(new ChatMessage(ChatRole.Assistant, "ok"));
-        ChatOptions? forwardedOptions = null;
-        using var selected = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, selectedOptions, _) =>
-            {
-                forwardedOptions = selectedOptions;
-                return Task.FromResult(expected);
-            },
-        };
-        RoutingContext? observedContext = null;
-        CancellationToken observedToken = default;
-        int selectionCount = 0;
-        using RoutingChatClient router = RoutingChatClient.Create((context, cancellationToken) =>
-        {
-            observedContext = context;
-            observedToken = cancellationToken;
-            selectionCount++;
-            context.ChatOptions!.ModelId = "selected";
-            return new(selected);
-        });
-
-        ChatResponse response = await router.GetResponseAsync(messages, options, cancellationSource.Token);
-
-        Assert.Same(expected, response);
-        Assert.Same(messages, observedContext!.Messages);
-        Assert.NotSame(options, observedContext.ChatOptions);
-        Assert.NotSame(observedContext.ChatOptions, forwardedOptions);
-        Assert.Equal("selected", observedContext.ChatOptions!.ModelId);
-        Assert.Equal("request", forwardedOptions!.ModelId);
-        Assert.Equal("request", options.ModelId);
-        Assert.Equal(cancellationSource.Token, observedToken);
-        Assert.Equal(1, selectionCount);
-    }
-
-    [Fact]
-    public async Task Create_DoesNotReselectAfterFailure()
-    {
-        var expected = new InvalidOperationException("failed");
-        using var selected = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) => throw expected,
-        };
-        int selectionCount = 0;
-        using RoutingChatClient router = RoutingChatClient.Create((_, _) =>
-        {
-            selectionCount++;
-            return new(selected);
-        });
-
-        InvalidOperationException actual = await Assert.ThrowsAsync<InvalidOperationException>(
-            () => router.GetResponseAsync([new(ChatRole.User, "hi")]));
-
-        Assert.Same(expected, actual);
-        Assert.Equal(1, selectionCount);
-    }
-
-    [Fact]
-    public async Task Create_NullSelectionThrowsForNonStreamingAndStreaming()
-    {
-        using RoutingChatClient router = RoutingChatClient.Create((_, _) => new((IChatClient)null!));
-
-        InvalidOperationException nonStreaming = await Assert.ThrowsAsync<InvalidOperationException>(
-            () => router.GetResponseAsync([new(ChatRole.User, "hi")]));
-        InvalidOperationException streaming = await Assert.ThrowsAsync<InvalidOperationException>(
-            () => CollectAsync(router.GetStreamingResponseAsync([new(ChatRole.User, "hi")])));
-
-        Assert.Contains("SelectClientAsync", nonStreaming.Message, StringComparison.Ordinal);
-        Assert.Contains("SelectClientAsync", streaming.Message, StringComparison.Ordinal);
-    }
-
-    [Fact]
-    public async Task Create_RejectsNullMessagesForNonStreamingAndStreaming()
-    {
-        using var selected = new TestChatClient();
-        using RoutingChatClient router = RoutingChatClient.Create((_, _) => new(selected));
-
-        await Assert.ThrowsAsync<ArgumentNullException>(() => router.GetResponseAsync(null!));
-        await Assert.ThrowsAsync<ArgumentNullException>(
-            () => CollectAsync(router.GetStreamingResponseAsync(null!)));
-    }
-
-    [Fact]
-    public async Task Create_DoesNotDisposeSelectedClient()
-    {
-        using var selected = new CountingDisposeClient();
-        using (RoutingChatClient router = RoutingChatClient.Create((_, _) => new(selected)))
-        {
-            _ = await router.GetResponseAsync([new(ChatRole.User, "hi")]);
-        }
-
-        Assert.Equal(0, selected.DisposeCount);
-    }
-
-    [Fact]
-    public void GetService_ReturnsSelfAndNullForUnknownOrKeyed()
-    {
-        using var client = new DelegatingTestRouter(_ => throw new NotSupportedException());
-
-        Assert.Same(client, client.GetService(typeof(DelegatingTestRouter)));
-        Assert.Same(client, client.GetService(typeof(RoutingChatClient)));
-        Assert.Same(client, client.GetService(typeof(IChatClient)));
-        Assert.Null(client.GetService(typeof(DelegatingTestRouter), serviceKey: "key"));
-        Assert.Null(client.GetService(typeof(string)));
-    }
-
     [Fact]
     public void MaximumAttemptsPerRequest_ValidatesValue()
     {
@@ -1322,540 +1188,6 @@ public class RoutingChatClientTests
         Assert.False(observed.ResponseCompleted);
     }
 
-    [Fact]
-    public void SemanticRouting_RejectsInvalidConfiguration()
-    {
-        using var client = new TestChatClient();
-        using var generator = new TestEmbeddingGenerator();
-        var profiles = new Dictionary<IChatClient, IReadOnlyList<string>>(ChatClientReferenceComparer.Instance)
-        {
-            [client] = ["profile"],
-        };
-
-        Assert.Throws<ArgumentNullException>(() =>
-            new SemanticRoutingChatClient(null!, profiles, client));
-        Assert.Throws<ArgumentNullException>(() =>
-            new SemanticRoutingChatClient(generator, null!, client));
-        Assert.Throws<ArgumentNullException>(() =>
-            new SemanticRoutingChatClient(generator, profiles, null!));
-        Assert.Throws<ArgumentException>(() =>
-            new SemanticRoutingChatClient(
-                generator,
-                new Dictionary<IChatClient, IReadOnlyList<string>>(),
-                client));
-        Assert.Throws<ArgumentException>(() =>
-            new SemanticRoutingChatClient(
-                generator,
-                new Dictionary<IChatClient, IReadOnlyList<string>>
-                {
-                    [client] = [],
-                },
-                client));
-        Assert.Throws<ArgumentException>(() =>
-            new SemanticRoutingChatClient(
-                generator,
-                new Dictionary<IChatClient, IReadOnlyList<string>>
-                {
-                    [client] = [" "],
-                },
-                client));
-        Assert.Throws<ArgumentOutOfRangeException>(() =>
-            new SemanticRoutingChatClient(generator, profiles, client, scoreThreshold: 1.1f));
-        Assert.Throws<ArgumentOutOfRangeException>(() =>
-            new SemanticRoutingChatClient(generator, profiles, client, topK: 0));
-        Assert.Throws<ArgumentOutOfRangeException>(() =>
-            new SemanticRoutingChatClient(
-                generator,
-                profiles,
-                client,
-                scoreAggregation: (SemanticRoutingChatClient.ScoreAggregation)(-1)));
-        Assert.Throws<ArgumentOutOfRangeException>(() =>
-            new SemanticRoutingChatClient(
-                generator,
-                profiles,
-                client,
-                scoreThreshold: 2.1f,
-                topK: 2,
-                scoreAggregation: SemanticRoutingChatClient.ScoreAggregation.Sum));
-    }
-
-    [Fact]
-    public async Task SemanticRouting_SelectsBestProfileAndCachesIndex()
-    {
-        var vectors = new Dictionary<string, float[]>
-        {
-            ["code profile"] = [1, 0],
-            ["writing profile"] = [0, 1],
-            ["debug this code"] = [1, 0],
-        };
-        int profileBatches = 0;
-        using var generator = new TestEmbeddingGenerator
-        {
-            GenerateAsyncCallback = (values, _, _) =>
-            {
-                string[] inputs = [.. values];
-                if (inputs.Length > 1)
-                {
-                    profileBatches++;
-                }
-
-                return Task.FromResult(new GeneratedEmbeddings<Embedding<float>>(
-                    [.. inputs.Select(input => new Embedding<float>(vectors[input]))]));
-            },
-        };
-        ChatResponse expected = new(new ChatMessage(ChatRole.Assistant, "code"));
-        using var code = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) => Task.FromResult(expected),
-        };
-        using var writing = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) =>
-                Task.FromResult(new ChatResponse(new ChatMessage(ChatRole.Assistant, "writing"))),
-        };
-        var profiles = new Dictionary<IChatClient, IReadOnlyList<string>>(ChatClientReferenceComparer.Instance)
-        {
-            [code] = ["code profile"],
-            [writing] = ["writing profile"],
-        };
-        using var router = new SemanticRoutingChatClient(
-            generator,
-            profiles,
-            defaultClient: writing,
-            leaveOpen: true);
-
-        ChatResponse first = await router.GetResponseAsync([new(ChatRole.User, "debug this code")]);
-        ChatResponse second = await router.GetResponseAsync([new(ChatRole.User, "debug this code")]);
-
-        Assert.Same(expected, first);
-        Assert.Same(expected, second);
-        Assert.Equal(1, profileBatches);
-    }
-
-    [Theory]
-    [InlineData(SemanticRoutingChatClient.ScoreAggregation.Mean, "code")]
-    [InlineData(SemanticRoutingChatClient.ScoreAggregation.Sum, "writing")]
-    public async Task SemanticRouting_AggregatesGlobalTopKByClient(
-        SemanticRoutingChatClient.ScoreAggregation scoreAggregation,
-        string expectedResponse)
-    {
-        var vectors = new Dictionary<string, float[]>
-        {
-            ["code"] = [1, 0],
-            ["writing one"] = [0.8f, 0.6f],
-            ["writing two"] = [0.8f, -0.6f],
-            ["query"] = [1, 0],
-        };
-        using var generator = new TestEmbeddingGenerator
-        {
-            GenerateAsyncCallback = (values, _, _) =>
-                Task.FromResult(new GeneratedEmbeddings<Embedding<float>>(
-                    [.. values.Select(input => new Embedding<float>(vectors[input]))])),
-        };
-        using var code = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) =>
-                Task.FromResult(new ChatResponse(new ChatMessage(ChatRole.Assistant, "code"))),
-        };
-        using var writing = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) =>
-                Task.FromResult(new ChatResponse(new ChatMessage(ChatRole.Assistant, "writing"))),
-        };
-        var profiles = new Dictionary<IChatClient, IReadOnlyList<string>>(ChatClientReferenceComparer.Instance)
-        {
-            [code] = ["code"],
-            [writing] = ["writing one", "writing two"],
-        };
-        using var router = new SemanticRoutingChatClient(
-            generator,
-            profiles,
-            defaultClient: code,
-            scoreThreshold: scoreAggregation == SemanticRoutingChatClient.ScoreAggregation.Sum ? 1.5f : 0.3f,
-            topK: 3,
-            scoreAggregation: scoreAggregation,
-            leaveOpen: true);
-
-        ChatResponse response = await router.GetResponseAsync([new(ChatRole.User, "query")]);
-
-        Assert.Equal(expectedResponse, response.Text);
-    }
-
-    [Fact]
-    public async Task SemanticRouting_UsesDefaultBelowThreshold()
-    {
-        var vectors = new Dictionary<string, float[]>
-        {
-            ["code profile"] = [1, 0],
-            ["unrelated query"] = [0, 1],
-        };
-        using var generator = new TestEmbeddingGenerator
-        {
-            GenerateAsyncCallback = (values, _, _) =>
-                Task.FromResult(new GeneratedEmbeddings<Embedding<float>>(
-                    [.. values.Select(input => new Embedding<float>(vectors[input]))])),
-        };
-        using var profiled = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) =>
-                Task.FromResult(new ChatResponse(new ChatMessage(ChatRole.Assistant, "profiled"))),
-        };
-        ChatResponse expected = new(new ChatMessage(ChatRole.Assistant, "default"));
-        using var defaultClient = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) => Task.FromResult(expected),
-        };
-        var profiles = new Dictionary<IChatClient, IReadOnlyList<string>>(ChatClientReferenceComparer.Instance)
-        {
-            [profiled] = ["code profile"],
-        };
-        using var router = new SemanticRoutingChatClient(
-            generator,
-            profiles,
-            defaultClient,
-            scoreThreshold: 0.5f,
-            leaveOpen: true);
-
-        ChatResponse response =
-            await router.GetResponseAsync([new(ChatRole.User, "unrelated query")]);
-
-        Assert.Same(expected, response);
-    }
-
-    [Fact]
-    public void SemanticRouting_DisposesOwnedResourcesOnce()
-    {
-#pragma warning disable CA2000 // Dispose objects before losing scope
-        var generator = new CountingEmbeddingGenerator();
-        var profiled = new CountingDisposeClient();
-        var defaultClient = new CountingDisposeClient();
-        var profiles = new Dictionary<IChatClient, IReadOnlyList<string>>(ChatClientReferenceComparer.Instance)
-        {
-            [profiled] = ["profile"],
-        };
-        var router = new SemanticRoutingChatClient(generator, profiles, defaultClient);
-#pragma warning restore CA2000
-
-        router.Dispose();
-        router.Dispose();
-
-        Assert.Equal(1, generator.DisposeCount);
-        Assert.Equal(1, profiled.DisposeCount);
-        Assert.Equal(1, defaultClient.DisposeCount);
-    }
-
-    [Fact]
-    public void OrderedFailover_RejectsMissingClients()
-    {
-        using var inner = new TestChatClient();
-
-        Assert.Throws<ArgumentNullException>(() => new OrderedFailoverChatClient(null!));
-        Assert.Throws<ArgumentException>(() => new OrderedFailoverChatClient([]));
-        Assert.Throws<ArgumentException>(() => new OrderedFailoverChatClient([inner, null!]));
-        Assert.Throws<ArgumentException>(() => new OrderedFailoverChatClient([inner, inner]));
-    }
-
-    [Fact]
-    public async Task OrderedFailover_TriesClientsInOrderAndSkipsFailures()
-    {
-        var calls = new List<string>();
-        using var first = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) =>
-            {
-                calls.Add("first");
-                throw new InvalidOperationException("first failed");
-            },
-        };
-        ChatResponse expected = new(new ChatMessage(ChatRole.Assistant, "ok"));
-        using var second = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) =>
-            {
-                calls.Add("second");
-                return Task.FromResult(expected);
-            },
-        };
-        using var client = new OrderedFailoverChatClient([first, second]);
-
-        ChatResponse response = await client.GetResponseAsync([new(ChatRole.User, "hi")]);
-
-        Assert.Same(expected, response);
-        Assert.Equal(["first", "second"], calls);
-    }
-
-    [Fact]
-    public async Task OrderedFailover_DistinguishesValueEqualClients()
-    {
-        ChatResponse expected = new(new ChatMessage(ChatRole.Assistant, "ok"));
-        using var first = new ValueEqualChatClient(
-            () => throw new InvalidOperationException("failed"));
-        using var second = new ValueEqualChatClient(
-            () => Task.FromResult(expected));
-        using var client = new OrderedFailoverChatClient([first, second]);
-
-        ChatResponse response = await client.GetResponseAsync([new(ChatRole.User, "hi")]);
-
-        Assert.Same(expected, response);
-    }
-
-    [Fact]
-    public async Task OrderedFailover_ExhaustionRethrowsLastFailure()
-    {
-        using var first = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) => throw new InvalidOperationException("first"),
-        };
-        using var second = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) => throw new InvalidOperationException("second"),
-        };
-        using var client = new OrderedFailoverChatClient([first, second]);
-
-        InvalidOperationException exception = await Assert.ThrowsAsync<InvalidOperationException>(
-            () => client.GetResponseAsync([new(ChatRole.User, "hi")]));
-
-        Assert.Equal("second", exception.Message);
-    }
-
-    [Fact]
-    public async Task OrderedFailover_StateIsScopedToOneRequest()
-    {
-        int firstCalls = 0;
-        using var first = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) =>
-            {
-                firstCalls++;
-                throw new InvalidOperationException("failed");
-            },
-        };
-        using var second = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) => Task.FromResult(new ChatResponse()),
-        };
-        using var client = new OrderedFailoverChatClient([first, second]);
-
-        _ = await client.GetResponseAsync([new(ChatRole.User, "one")]);
-        _ = await client.GetResponseAsync([new(ChatRole.User, "two")]);
-
-        Assert.Equal(2, firstCalls);
-    }
-
-    [Fact]
-    public async Task OrderedFailover_ConcurrentRequestsHaveIndependentState()
-    {
-        const int RequestCount = 10;
-        int firstCalls = 0;
-        int secondCalls = 0;
-        var allStarted = new TaskCompletionSource<bool>(TaskCreationOptions.RunContinuationsAsynchronously);
-        using var first = new TestChatClient
-        {
-            GetResponseAsyncCallback = async (_, _, _) =>
-            {
-                if (Interlocked.Increment(ref firstCalls) == RequestCount)
-                {
-                    allStarted.SetResult(true);
-                }
-
-                await allStarted.Task;
-                throw new InvalidOperationException("failed");
-            },
-        };
-        using var second = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) =>
-            {
-                Interlocked.Increment(ref secondCalls);
-                return Task.FromResult(new ChatResponse());
-            },
-        };
-        using var client = new OrderedFailoverChatClient([first, second]);
-
-        Task<ChatResponse>[] requests =
-        [
-            .. Enumerable.Range(0, RequestCount).Select(index =>
-                client.GetResponseAsync([new(ChatRole.User, index.ToString())])),
-        ];
-        _ = await Task.WhenAll(requests);
-
-        Assert.Equal(RequestCount, firstCalls);
-        Assert.Equal(RequestCount, secondCalls);
-    }
-
-    [Fact]
-    public async Task OrderedFailover_DoesNotRetainStateWhileStreaming()
-    {
-        using var first = new TestChatClient
-        {
-            GetStreamingResponseAsyncCallback = (_, _, _) => ThrowingStream("failed"),
-        };
-        using var second = new TestChatClient
-        {
-            GetStreamingResponseAsyncCallback = (_, _, _) => YieldUpdates("first", "second"),
-        };
-        using var client = new OrderedFailoverChatClient([first, second]);
-        await using IAsyncEnumerator<ChatResponseUpdate> enumerator =
-            client.GetStreamingResponseAsync([new(ChatRole.User, "hi")]).GetAsyncEnumerator();
-
-        Assert.True(await enumerator.MoveNextAsync());
-        Assert.Equal("first", enumerator.Current.Text);
-        Assert.Equal(0, GetOrderedFailoverRequestStateCount(client));
-    }
-
-    [Fact]
-    public async Task OrderedFailover_SnapshotsConfiguredClients()
-    {
-        ChatResponse expected = new(new ChatMessage(ChatRole.Assistant, "ok"));
-        using var inner = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) => Task.FromResult(expected),
-        };
-        var clients = new List<IChatClient> { inner };
-        using var failover = new OrderedFailoverChatClient(clients, leaveOpen: true);
-        clients.Clear();
-
-        ChatResponse response = await failover.GetResponseAsync([new(ChatRole.User, "hi")]);
-
-        Assert.Same(expected, response);
-    }
-
-    [Fact]
-    public async Task OrderedFailover_UsesEachConfiguredClient()
-    {
-        var seenOptions = new List<(string? modelId, string? instructions)>();
-        using var firstInner = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, options, _) =>
-            {
-                seenOptions.Add((options?.ModelId, options?.Instructions));
-                throw new InvalidOperationException("failed");
-            },
-        };
-        using var first = new ConfigureOptionsChatClient(
-            firstInner,
-            options => options.ModelId = "first");
-        using var secondInner = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, options, _) =>
-            {
-                seenOptions.Add((options?.ModelId, options?.Instructions));
-                return Task.FromResult(new ChatResponse());
-            },
-        };
-        using var second = new ConfigureOptionsChatClient(
-            secondInner,
-            options => options.ModelId = "second");
-        using var client = new OrderedFailoverChatClient(
-            [first, second],
-            leaveOpen: true);
-
-        _ = await client.GetResponseAsync(
-            [new(ChatRole.User, "hi")],
-            new ChatOptions
-            {
-                Instructions = "caller",
-                ModelId = "request",
-            });
-
-        Assert.Equal(
-            [("first", "caller"), ("second", "caller")],
-            seenOptions);
-    }
-
-    [Fact]
-    public async Task OrderedFailover_StreamingFallsBackBeforeFirstUpdate()
-    {
-        using var first = new TestChatClient
-        {
-            GetStreamingResponseAsyncCallback = (_, _, _) => ThrowingStream("failed"),
-        };
-        using var second = new TestChatClient
-        {
-            GetStreamingResponseAsyncCallback = (_, _, _) => YieldUpdates("ok"),
-        };
-        using var client = new OrderedFailoverChatClient([first, second]);
-
-        List<ChatResponseUpdate> updates =
-            await CollectAsync(client.GetStreamingResponseAsync([new(ChatRole.User, "hi")]));
-
-        Assert.Equal("ok", Assert.Single(updates).Text);
-    }
-
-    [Fact]
-    public async Task OrderedFailover_CancellationDoesNotFallback()
-    {
-        using var cancellationSource = new CancellationTokenSource();
-        cancellationSource.Cancel();
-        bool secondCalled = false;
-        using var first = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) => throw new InvalidOperationException("failed"),
-        };
-        using var second = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) =>
-            {
-                secondCalled = true;
-                return Task.FromResult(new ChatResponse());
-            },
-        };
-        using var client = new OrderedFailoverChatClient([first, second]);
-
-        await Assert.ThrowsAnyAsync<OperationCanceledException>(
-            () => client.GetResponseAsync(
-                [new(ChatRole.User, "hi")],
-                cancellationToken: cancellationSource.Token));
-
-        Assert.False(secondCalled);
-    }
-
-    [Fact]
-    public void OrderedFailover_DisposesEachClientOnce()
-    {
-#pragma warning disable CA2000 // Dispose objects before losing scope
-        var shared = new CountingDisposeClient();
-        var other = new CountingDisposeClient();
-        var client = new OrderedFailoverChatClient([shared, other]);
-#pragma warning restore CA2000
-
-        client.Dispose();
-        client.Dispose();
-
-        Assert.Equal(1, shared.DisposeCount);
-        Assert.Equal(1, other.DisposeCount);
-    }
-
-    [Fact]
-    public void OrderedFailover_LeaveOpenDoesNotDisposeInnerClients()
-    {
-#pragma warning disable CA2000 // Dispose objects before losing scope
-        var inner = new CountingDisposeClient();
-        var client = new OrderedFailoverChatClient([inner], leaveOpen: true);
-#pragma warning restore CA2000
-
-        client.Dispose();
-
-        Assert.Equal(0, inner.DisposeCount);
-        inner.Dispose();
-    }
-
-    [Fact]
-    public async Task NestedRouters_ReturnLeafResponse()
-    {
-        ChatResponse expected = new(new ChatMessage(ChatRole.Assistant, "ok"));
-        using var leaf = new TestChatClient
-        {
-            GetResponseAsyncCallback = (_, _, _) => Task.FromResult(expected),
-        };
-        using var inner = new OrderedFailoverChatClient([leaf]);
-        using var outer = new OrderedFailoverChatClient([inner]);
-
-        ChatResponse response = await outer.GetResponseAsync([new(ChatRole.User, "hi")]);
-
-        Assert.Same(expected, response);
-    }
-
     private static async Task<List<ChatResponseUpdate>> CollectAsync(
         IAsyncEnumerable<ChatResponseUpdate> updates)
     {
@@ -1872,14 +1204,6 @@ public class RoutingChatClientTests
     {
         await using IAsyncEnumerator<ChatResponseUpdate> enumerator = updates.GetAsyncEnumerator();
         Assert.True(await enumerator.MoveNextAsync());
-    }
-
-    private static int GetOrderedFailoverRequestStateCount(OrderedFailoverChatClient client)
-    {
-        object requestStates = typeof(OrderedFailoverChatClient)
-            .GetField("_requestStates", System.Reflection.BindingFlags.Instance | System.Reflection.BindingFlags.NonPublic)!
-            .GetValue(client)!;
-        return (int)requestStates.GetType().GetProperty("Count")!.GetValue(requestStates)!;
     }
 
     private static async IAsyncEnumerable<ChatResponseUpdate> YieldUpdates(params string[] texts)
@@ -2062,80 +1386,5 @@ public class RoutingChatClientTests
                 return default;
             }
         }
-    }
-
-    private sealed class CountingDisposeClient : IChatClient
-    {
-        public int DisposeCount { get; private set; }
-
-        public Task<ChatResponse> GetResponseAsync(
-            IEnumerable<ChatMessage> messages,
-            ChatOptions? options = null,
-            CancellationToken cancellationToken = default) =>
-            Task.FromResult(new ChatResponse());
-
-        public IAsyncEnumerable<ChatResponseUpdate> GetStreamingResponseAsync(
-            IEnumerable<ChatMessage> messages,
-            ChatOptions? options = null,
-            CancellationToken cancellationToken = default) =>
-            throw new NotSupportedException();
-
-        public object? GetService(Type serviceType, object? serviceKey = null) => null;
-
-        public void Dispose() => DisposeCount++;
-
-        public override bool Equals(object? obj) => obj is CountingDisposeClient;
-
-        public override int GetHashCode() => 0;
-    }
-
-    private sealed class CountingEmbeddingGenerator :
-        IEmbeddingGenerator<string, Embedding<float>>
-    {
-        public int DisposeCount { get; private set; }
-
-        public Task<GeneratedEmbeddings<Embedding<float>>> GenerateAsync(
-            IEnumerable<string> values,
-            EmbeddingGenerationOptions? options = null,
-            CancellationToken cancellationToken = default) =>
-            throw new NotSupportedException();
-
-        public object? GetService(Type serviceType, object? serviceKey = null) => null;
-
-        public void Dispose() => DisposeCount++;
-    }
-
-    private sealed class ChatClientReferenceComparer : IEqualityComparer<IChatClient>
-    {
-        public static ChatClientReferenceComparer Instance { get; } = new();
-
-        public bool Equals(IChatClient? x, IChatClient? y) => ReferenceEquals(x, y);
-
-        public int GetHashCode(IChatClient obj) => RuntimeHelpers.GetHashCode(obj);
-    }
-
-    private sealed class ValueEqualChatClient(Func<Task<ChatResponse>> getResponse) : IChatClient
-    {
-        public Task<ChatResponse> GetResponseAsync(
-            IEnumerable<ChatMessage> messages,
-            ChatOptions? options = null,
-            CancellationToken cancellationToken = default) =>
-            getResponse();
-
-        public IAsyncEnumerable<ChatResponseUpdate> GetStreamingResponseAsync(
-            IEnumerable<ChatMessage> messages,
-            ChatOptions? options = null,
-            CancellationToken cancellationToken = default) =>
-            throw new NotSupportedException();
-
-        public object? GetService(Type serviceType, object? serviceKey = null) => null;
-
-        public void Dispose()
-        {
-        }
-
-        public override bool Equals(object? obj) => obj is ValueEqualChatClient;
-
-        public override int GetHashCode() => 0;
     }
 }
